@@ -4,67 +4,50 @@
   "private": true,
   "main": "./src/index.ts",
   "dependencies": {
-    "bootstrap": "^5.1.3",
+    "@tkcrm/ui": "^0.0.4",
+    "classnames": "^2.3.1",
     "history": "^5.3.0",
-    "i18next": "^21.6.13",
-    "i18next-browser-languagedetector": "^6.1.3",
-    "i18next-xhr-backend": "^3.2.2",
-    "react": "^17.0.2",
-    "react-i18next": "^11.15.5",
-    "react-router": "^6.2.2",
-    "react-router-dom": "^6.2.2"
+    "i18next": "^21.6.16",
+    "i18next-browser-languagedetector": "^6.1.4",
+    "i18next-http-backend": "^1.4.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-i18next": "^11.16.7",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0"
   },
   "devDependencies": {
-    "@types/classnames": "^2.3.1",
-    "@types/qs": "^6.9.7",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.13",
+    "@tailwindcss/forms": "^0.5.0",
+    "@types/node": "^17.0.25",
+    "@types/react": "^18.0.6",
+    "@types/react-dom": "^18.0.2",
     "@types/react-router": "^5.1.18",
     "@types/react-router-dom": "^5.3.3",
-    "@typescript-eslint/eslint-plugin": "^5.13.0",
-    "@typescript-eslint/parser": "^5.13.0",
-    "@webpack-cli/serve": "^1.6.1",
-    "clean-webpack-plugin": "^4.0.0",
-    "copy-webpack-plugin": "^10.2.4",
-    "css-loader": "^6.6.0",
-    "css-minimizer-webpack-plugin": "^3.4.1",
-    "dotenv": "^16.0.0",
-    "dotenv-expand": "^8.0.1",
-    "eslint": "^8.10.0",
+    "@vitejs/plugin-react": "^1.3.1",
+    "autoprefixer": "^10.4.4",
+    "cssnano": "^5.1.7",
+    "eslint": "^8.13.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.29.3",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "eslint-plugin-unicorn": "^41.0.0",
+    "eslint-plugin-react": "^7.29.4",
+    "eslint-plugin-react-hooks": "^4.4.0",
+    "eslint-plugin-unicorn": "^42.0.0",
     "eslint-plugin-you-dont-need-lodash-underscore": "^6.12.0",
-    "eslint-webpack-plugin": "^3.1.1",
-    "html-webpack-plugin": "^5.5.0",
-    "mini-css-extract-plugin": "^2.6.0",
-    "mini-svg-data-uri": "^1.4.3",
-    "prettier": "2.5.1",
-    "sass": "^1.49.9",
-    "sass-loader": "^12.6.0",
-    "source-map-loader": "^3.0.1",
-    "style-loader": "^3.3.1",
-    "ts-import-plugin": "^2.0.0",
-    "ts-loader": "^9.2.7",
-    "tsconfig-paths-webpack-plugin": "^3.5.2",
-    "tslint": "^6.1.3",
-    "typescript": "^4.6.2",
-    "webpack": "^5.70.0",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "4.7.4",
-    "webpack-manifest-plugin": "^4.1.1",
-    "webpack-merge": "^5.8.0",
-    "workbox-webpack-plugin": "^6.5.1"
+    "postcss-import": "^14.1.0",
+    "prettier": "^2.6.2",
+    "prettier-plugin-tailwindcss": "^0.1.10",
+    "tailwindcss": "^3.0.24",
+    "typescript": "^4.6.3",
+    "vite": "^2.9.5"
   },
   "scripts": {
-    "start": "webpack-cli serve --env NODE_ENV=development --mode development --config configs/webpack/webpack.dev.js",
-    "build": "webpack --mode production --config configs/webpack/webpack.prod.js",
+    "start": "vite",
+    "build": "tsc && vite build",
     "lint": "eslint src/** --ignore-pattern '/src/assets/*'",
-    "docker:build": "docker build . -t frontend-boilerplate-test",
-    "docker:start": "docker run --name frontend-boilerplate-test -it -d -p 8080:8080 --rm frontend-boilerplate-test"
+    "docker:build": "npm run build && docker build . -t frontend-boilerplate",
+    "docker:push": "docker push . -t frontend-boilerplate",
+    "docker:start": "docker run --name frontend-boilerplate -it -d -p 8080:8080 --rm frontend-boilerplate"
   },
   "browserslist": {
     "production": [
